仓库添加修改的  去重除己 仓库删除的 关联数据删除

+++ module/product/store.action.php
@@ -109,12 +109,13 @@ class storeAction extends adminBaseAction
 		if(empty($ids)){
 			$this->error('操作有误');
 		}
-		$admin=sget('remove','s');//非空是删除管理员，来自storeadmin.list.html
+		$admin=sget('remove','s');//如果非空代表是删除管理员，来自storeadmin.list.html
 		if(!empty($admin)){
 			$result=$this->db->model('store_admin')->where("id in ($ids)")->delete();
 		}
 		$result=$this->db->where("id in ($ids)")->delete();
-		if($result){
+		$result2=$this->db->model('store_admin')->where("store_id in ($ids)")->delete();//删除仓库，关联删除绑定联系人
+		if($result && $result2){
 			$cache=cache::startMemcache();
 			$cache->delete('store');
 			$this->success('操作成功');
